refactor: 주문 서비스 Money 값 객체 적용

+++ coupon-server/src/test/java/com/drf/coupon/calculator/CartCouponCalculatorTest.java
@@ -72,7 +72,7 @@ class CartCouponCalculatorTest {
                 .build();
     }
 
-    private InternalCartCouponItemRequest item(long cartItemId, long productId, int price, int quantity, List<Long> categoryPath) {
+    private InternalCartCouponItemRequest item(long cartItemId, long productId, long price, int quantity, List<Long> categoryPath) {
         return new InternalCartCouponItemRequest(cartItemId, productId, price, quantity, categoryPath);
     }
