+++ order-server/src/main/java/com/drf/order/client/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
 package com.drf.order.client.dto.response;
 
-public record ProductResponse(long id, String name, int price, int stock, String status) {
+public record ProductResponse(long id, String name, long price, int stock, String status) {
 }
