+++ order-server/src/main/java/com/drf/order/model/response/OrderCreateResponse.java
@@ -3,5 +3,5 @@ package com.drf.order.model.response;
 import lombok.Builder;
 
 @Builder
-public record OrderCreateResponse(long orderId, String orderNo, String status, int finalAmount) {
+public record OrderCreateResponse(long orderId, String orderNo, String status, long finalAmount) {
 }
